requirements.txt
streamlit
transformers
torch
pandas
matplotlib
python-docx